fix web input focus

+++ src/models/personality.py
@@ -41,7 +41,7 @@ def get_llm_model_config() -> Dict[str, Any]:
         "temperature": float(os.getenv("OLLAMA_TEMPERATURE", 1.0)),
         "num_predict": int(os.getenv("OLLAMA_NUM_PREDICT", 40)),
         "repeat_penalty": float(os.getenv("OLLAMA_REPEAT_PENALTY", 1.2)),
-        "top_k": float(os.getenv("OLLAMA_TOP_K", 40)),
+        "top_k": int(os.getenv("OLLAMA_TOP_K", 40)),
         "top_p": float(os.getenv("OLLAMA_TOP_P", 0.9)),
         "stop": ["User:", "Pip:"],
     }
